Vendor RSpec::Matchers::BuiltIn::BaseMatcher.
This is in aid of moving RSpec Rails off RSpec expectations private API.

Going forward, it would be cool to refactor this base matcher to not be
a total and direct copy of RSpec Expectations' implementation, and
rather an implementation of the matcher API that rails needs, but for
right now this was the fastest route to green.

Going forward in to RSpec 4, it's good to flag that this might cause breakage,
given that the internal matcher API might change. I'd say that the matcher API
is mostly stable enough at this point, so I'm not hugely concerned.
@myronmarston, @jonrowe, @benoittgt do you predict the RSpec matchers API
changing drastically enough in RSpec 4 to cause this to be problematic?

One thing that I'd note here is that I think it'd be cool if we had a really
good step by step guide for implementing the matcher protocol from scratch, as
opposed to using custom matchers or inheriting from base matcher. Right now
there's no obvious human readable documentation for that. Maybe that's
something I could find a new contributor to do?